requirements.txt
streamlit>=1.28.0
biopython>=1.81
stmol>=0.0.9
reportlab>=4.0.0
fpdf2>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
